kernel: move thread_stack.h to include/kernel
Move this internal header file into kernel/ alongside other thread
related headers.
